+++ EventSourcing.Domain/Account/WithdrawalEvent.cs
@@ -5,19 +5,19 @@ namespace EventSourcing.Domain.Account
 {
     public class WithdrawalEvent : DomainEventBase<Guid>
     {
-        public double Amount { get; private set; }
+        public decimal Amount { get; private set; }
 
         public WithdrawalEvent()
         {
 
         }
 
-        public WithdrawalEvent(Guid aggregateId, double amount) : base(aggregateId)
+        public WithdrawalEvent(Guid aggregateId, decimal amount) : base(aggregateId)
         {
             Amount = amount;
         }
 
-        public WithdrawalEvent(Guid aggregateId, long aggregateVersion, double amount) : base(aggregateId, aggregateVersion)
+        public WithdrawalEvent(Guid aggregateId, long aggregateVersion, decimal amount) : base(aggregateId, aggregateVersion)
         {
             Amount = amount;
         }
